Separate core components into other files
