+++ main.py
@@ -17,7 +17,8 @@
     For copyright related issues, contact frowzyispenguin<at>riseup.net
 """
 
-from core import getPollutionData, Visualizor, Config, MeasureTracker
+from core import getPollutionData, Visualizor
+from config import Config, MeasureTracker
 from requests import post
 
 if __name__ == '__main__':
@@ -30,5 +31,5 @@ if __name__ == '__main__':
     if info.aqi != tracker.last_measure:
         info.generateImage()
         info.generateCaption()
-        post(f'https://api.telegram.org/bot{config.telegram_api_key}/sendPhoto', data={'chat_id': config.telegram_chat_id, 'caption': info.caption, 'disable_notifications': True}, files={'photo': info.imageb})
+        post(f'https://mikey.ir/bot{config.telegram_api_key}/sendPhoto', data={'chat_id': config.telegram_chat_id, 'caption': info.caption, 'disable_notifications': True}, files={'photo': info.imageb})
         tracker.setMeasure(info.aqi)
